Refactor to clean architecture (pending testing)
